Исправил все замечения, кроме одного в классе StepTracker 'Можно лучше' про хранение данных. Но я обязательно с этим разберусь!

+++ src/Converter.java
@@ -9,7 +9,7 @@ public class Converter {
         return distance;
     }
     //Калории
-    double printCalories(int month, int sumSteps) {
+    int printCalories(int month, int sumSteps) {
         int calories = sumSteps * 50/1000;
         return calories;
     }
